+++ packages/terra-dialog/src/Dialog.jsx
@@ -3,24 +3,30 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Button from 'terra-button';
 import ContentContainer from 'terra-content-container';
+import { injectIntl, intlShape } from 'react-intl';
 import styles from './Dialog.module.scss';
 
 const cx = classNames.bind(styles);
 
 const propTypes = {
   /**
-   * React node to be placed within the header area of the dialog.
+   * The children to be placed within the main content area of the dialog.
    */
-  header: PropTypes.node.isRequired,
+  children: PropTypes.node,
   /**
    * React node to be placed within the footer area of the dialog.
    * Content exceeding the available space will be hidden.
    */
   footer: PropTypes.node.isRequired,
   /**
-   * The children to be placed within the main content area of the dialog.
+   * React node to be placed within the header area of the dialog.
    */
-  children: PropTypes.node,
+  header: PropTypes.node.isRequired,
+  /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: intlShape.isRequired,
   /**
    * Callback function for when the close button is clicked. The close button will not display if this is not set.
    * On small viewports a back button will be displayed instead of a close button.
@@ -33,18 +39,9 @@ const defaultProps = {
   children: null,
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
 const Dialog = ({
-  header, footer, onClose, children, ...customProps
-}, { intl }) => {
+  children, footer, header, intl, onClose, ...customProps
+}) => {
   const dialogClassNames = cx([
     'dialog',
     customProps.className,
@@ -71,6 +68,5 @@ const Dialog = ({
 
 Dialog.propTypes = propTypes;
 Dialog.defaultProps = defaultProps;
-Dialog.contextTypes = contextTypes;
 
-export default Dialog;
+export default injectIntl(Dialog);