+++ packages/terra-form-field/src/Field.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { injectIntl, intlShape } from 'react-intl';
 import classNames from 'classnames/bind';
 import IconError from 'terra-icon/lib/icon/IconError';
 import styles from './Field.module.scss';
@@ -31,6 +32,11 @@ const propTypes = {
    * The htmlFor attribute on the field label.
    */
   htmlFor: PropTypes.string,
+  /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: intlShape.isRequired,
   /**
    * Whether or not the field is invalid.
    */
@@ -88,21 +94,12 @@ const defaultProps = {
   showOptional: false,
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
 const hasWhiteSpace = s => /\s/g.test(s);
 // Detect IE 10 or IE 11
 // TODO - Delete detection logic when we drop support for IE
 const isIE = () => (window.navigator.userAgent.indexOf('Trident/6.0') > -1 || window.navigator.userAgent.indexOf('Trident/7.0') > -1);
 
-const Field = (props, { intl }) => {
+const Field = (props) => {
   const {
     children,
     error,
@@ -110,6 +107,7 @@ const Field = (props, { intl }) => {
     help,
     hideRequired,
     htmlFor,
+    intl,
     isInvalid,
     isInline,
     isLabelHidden,
@@ -203,6 +201,5 @@ const Field = (props, { intl }) => {
 
 Field.propTypes = propTypes;
 Field.defaultProps = defaultProps;
-Field.contextTypes = contextTypes;
 
-export default Field;
+export default injectIntl(Field);