+++ packages/terra-action-header/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
+
+2.20.0 - (July 16, 2019)
+------------------
 ### Added
 * Added documentation on required peerDependencies